requirements.txt
# Core Dependencies
python-binance==1.0.17
ccxt==4.1.69
web3==6.11.3
websockets==11.0.3
aiohttp==3.9.1

# FastAPI & Web Framework and API
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
websockets==12.0
prometheus-client==0.19.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# AI/ML Dependencies
tensorflow==2.15.0
scikit-learn==1.3.2
torch==2.1.2

# Technical Analysis
ta==0.11.0
pandas-ta==0.3.14b0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
timescaledb==0.2.0
redis==5.0.1

# System Monitoring
psutil==5.9.6
prometheus-client==0.19.0
grafana-api==1.0.3

# Social Media APIs
tweepy==4.14.0
praw==7.7.1
python-telegram-bot==20.7
discord.py==2.3.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# API & Communication
tweepy==4.14.0
python-telegram-bot==20.7
discord.py==2.3.2

# Blockchain Specific
eth-account==0.10.0
eth-utils==2.3.1

# Performance
uvloop==0.19.0
cython==3.0.7

# Additional Tools
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
aiofiles==23.2.1

# Security & Authentication  
cryptography==41.0.7
bcrypt==4.1.2

# Enhanced Dependencies for New Components
dataclasses-json==0.6.3
python-multipart==0.0.6
email-validator==2.1.0
